+++ assets/router/index.js
@@ -1,10 +1,12 @@
 import { createRouter, createWebHistory } from 'vue-router';
-import DashboardView from '../vue/DashboardView.vue'
+import DashboardView from '../vue/views/DashboardView.vue'
+import SubscribersView from '../vue/views/SubscribersView.vue'
 
 export const router = createRouter({
     history: createWebHistory(),
     routes: [
-        { path: '/', name: 'dashboard', component: DashboardView },
+        { path: '/dashboard', name: 'dashboard', component: DashboardView },
+        { path: '/subscribers', name: 'subscribers', component: SubscribersView },
         { path: '/:pathMatch(.*)*', redirect: '/' },
     ],
 });
