+++ assets/vue/components/dashboard/CampaignsTable.vue
